refactor: Duration 파일 이동
