+++ src/main/java/com/programmers/springbootbasic/domain/voucher/FixedAmountVoucher.java
@@ -1,5 +1,7 @@
 package com.programmers.springbootbasic.domain.voucher;
 
+import com.programmers.springbootbasic.domain.model.Duration;
+
 import java.util.UUID;
 
 public class FixedAmountVoucher extends Voucher {
