+++ src/main/java/com/programmers/springbootbasic/domain/voucher/PercentDiscountVoucher.java
@@ -1,5 +1,7 @@
 package com.programmers.springbootbasic.domain.voucher;
 
+import com.programmers.springbootbasic.domain.model.Duration;
+
 import java.util.UUID;
 
 public class PercentDiscountVoucher extends Voucher {
