+++ src/main/java/com/programmers/springbootbasic/service/dto/Voucher/VoucherResponse.java
@@ -1,7 +1,7 @@
 package com.programmers.springbootbasic.service.dto.Voucher;
 
 
-import com.programmers.springbootbasic.domain.voucher.Duration;
+import com.programmers.springbootbasic.domain.model.Duration;
 
 public record VoucherResponse(
         String name,
