+++ src/test/java/com/programmers/springbootbasic/domain/voucher/VoucherTest.java
@@ -1,5 +1,6 @@
 package com.programmers.springbootbasic.domain.voucher;
 
+import com.programmers.springbootbasic.domain.model.Duration;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
